Fixed RedPackage

+++ WolvenKit.Common/RED4/CR2W/JSON/PrimitiveConverter.cs
@@ -1196,7 +1196,7 @@ public class RedPackageConverter : CustomRedConverter<RedPackage>
                         throw new JsonException();
                     }
 
-                    result.Version = reader.GetUInt16();
+                    result.Version = reader.GetByte();
                     break;
                 }
